54 fix 후기 입찰 경매물품 배송 fk 설정

+++ market/src/main/java/com/charly/market/auction_item/model/dto/AuctionItemResponse.java
@@ -1,5 +1,7 @@
 package com.charly.market.auction_item.model.dto;
 
+import com.charly.market.user.model.entity.User;
+
 import java.time.LocalDateTime;
 
 
@@ -10,7 +12,7 @@ public record AuctionItemResponse(
          long bidUnit,
          LocalDateTime auctionStartTime,
          LocalDateTime auctionEndTime,
-         int userId
+         Long userId
 ){
